src/test/java/ml/pedidos/api/controller/ControllerValidadorDeRequestsTest.java
package ml.pedidos.api.controller;

import ml.pedidos.api.response.ErrorResponse;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.BindingResult;
import org.springframework.validation.FieldError;
import org.springframework.web.bind.MethodArgumentNotValidException;

import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
public class ControllerValidadorDeRequestsTest {

    @InjectMocks
    private ControllerValidadorDeRequests controllerValidadorDeRequests;

    @Test
    public void testHandleValidationExceptions() {
        MethodArgumentNotValidException ex = mock(MethodArgumentNotValidException.class);
        BindingResult bindingResult = mock(BindingResult.class);
        FieldError fieldError1 = mock(FieldError.class);
        FieldError fieldError2 = mock(FieldError.class);

        when(ex.getBindingResult()).thenReturn(bindingResult);
        when(bindingResult.getFieldErrors()).thenReturn(List.of(fieldError1, fieldError2));
        when(fieldError1.getDefaultMessage()).thenReturn("Erro 1");
        when(fieldError2.getDefaultMessage()).thenReturn("Erro 2");

        ResponseEntity<List<ErrorResponse>> response = controllerValidadorDeRequests.handleValidationExceptions(ex);

        assertEquals(HttpStatus.BAD_REQUEST, response.getStatusCode());
        assertEquals(2, response.getBody().size());
        assertEquals("Erro 1", response.getBody().get(0).getMensagem());
        assertEquals("Erro 2", response.getBody().get(1).getMensagem());
    }
}